Make stack traces more succinct.
Currently stack traces are quite verbose. For each form on the stack, a whole
stack trace is shown. So if an exception occurs on the 5th form on the stack,
you get 5 stack traces.

This change reduces the number of stack traces to 2: the original exception,
and an exception showing the form stack.